Refactor test utils function into packages
