+++ src/test/kotlin/io/github/raeperd/realworldspringbootkotlin/ArticleFavoriteIntegrationTest.kt
@@ -1,7 +1,7 @@
 package io.github.raeperd.realworldspringbootkotlin
 
-import io.github.raeperd.realworldspringbootkotlin.util.JpaDatabaseCleanerExtension
-import io.github.raeperd.realworldspringbootkotlin.util.andReturnResponseBody
+import io.github.raeperd.realworldspringbootkotlin.util.junit.JpaDatabaseCleanerExtension
+import io.github.raeperd.realworldspringbootkotlin.util.spring.andReturnResponseBody
 import io.github.raeperd.realworldspringbootkotlin.web.ArticleModel
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
