+++ src/test/kotlin/io/github/raeperd/realworldspringbootkotlin/util/jackson/SingletonObjectMapper.kt
@@ -1,16 +1,11 @@
-package io.github.raeperd.realworldspringbootkotlin.util
+package io.github.raeperd.realworldspringbootkotlin.util.jackson
 
-import com.fasterxml.jackson.annotation.JsonInclude.Include
+import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.KotlinFeature
 import com.fasterxml.jackson.module.kotlin.KotlinModule
-import com.fasterxml.jackson.module.kotlin.readValue
-import org.hamcrest.Matchers
-import org.springframework.test.web.servlet.MockMvcResultMatchersDsl
-import org.springframework.test.web.servlet.ResultActionsDsl
-import org.springframework.test.web.servlet.result.ContentResultMatchersDsl
 
 object SingletonObjectMapper : ObjectMapper() {
     init {
@@ -26,21 +21,8 @@ object SingletonObjectMapper : ObjectMapper() {
         )
         registerModule(JavaTimeModule())
         configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
-        setSerializationInclusion(Include.NON_NULL)
+        setSerializationInclusion(JsonInclude.Include.NON_NULL)
     }
 }
 
 fun <T> T.toJson(): String = SingletonObjectMapper.writeValueAsString(this)
-
-fun MockMvcResultMatchersDsl.notEmptyErrorResponse() {
-    return jsonPath("errors.body", Matchers.not(emptyList<String>()))
-}
-
-fun <T> ContentResultMatchersDsl.responseJson(dto: T) {
-    json(dto.toJson())
-}
-
-inline fun <reified T> ResultActionsDsl.andReturnResponseBody(): T {
-    return andReturn().response.contentAsString
-        .let { SingletonObjectMapper.readValue(it) }
-}
